# contracts/openapi/portfolio-api.json
{
  "components": {
    "schemas": {
...
      "CashResponse": {
        "properties": {
          "accountName": {
            "type": "string"
          },
          "accountNumberMasked": {
            "type": "string"
          },
          "currentValue": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "rowNumber": {
            "format": "int32",
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "ClassificationRequest": {
        "properties": {
          "classification": {
            "minLength": 1,
            "type": "string"
          },
          "expectedVersion": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": ["classification"],
        "type": "object"
      },
      "ClassificationSuggestionResponse": {
        "properties": {
          "assetType": {
            "type": "string"
          },
          "blocked": {
            "type": "boolean"
          },
          "classification": {
            "type": "string"
          },
          "confirmationRequired": {
            "type": "boolean"
          },
          "positionId": {
            "format": "uuid",
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "DipPreviewRequest": {
        "properties": {
          "aboveEma20": {
            "type": "boolean"
          },
          "breadthImproving": {
            "type": "boolean"
          },
          "breadthOversoldScore": {
            "format": "double",
            "type": "number"
          },
          "breakout5Day": {
            "type": "boolean"
          },
          "completeData": {
            "type": "boolean"
          },
          "completedTranches": {
            "format": "int32",
            "maximum": 4,
            "minimum": 0,
            "type": "integer"
          },
          "creditStable": {
            "type": "boolean"
          },
          "creditStressScore": {
            "format": "double",
            "type": "number"
          },
          "drawdownScore": {
            "format": "double",
            "type": "number"
          },
          "emergencyCashProtected": {
            "type": "boolean"
          },
          "marketDriven": {
            "type": "boolean"
          },
          "portfolioDrawdown": {
            "minimum": 0,
            "type": "number"
          },
          "rsiCross40": {
            "type": "boolean"
          },
          "tradingDaysSinceLastTranche": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "trendContextScore": {
            "format": "double",
            "type": "number"
          },
          "vixFalling": {
            "type": "boolean"
          },
          "vixPercentileScore": {
            "format": "double",
            "type": "number"
          },
          "volTermScore": {
            "format": "double",
            "type": "number"
          }
        },
        "required": ["portfolioDrawdown"],
        "type": "object"
      },
      "DipPreviewResponse": {
        "properties": {
          "action": {
            "type": "string"
          },
          "debug": {
            "type": "boolean"
          },
          "executionSubmitted": {
            "type": "boolean"
          },
          "reserveFraction": {
            "type": "string"
          },
          "ruleIds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "setupScore": {
            "format": "double",
            "type": "number"
          },
          "trancheNumber": {
            "format": "int32",
            "type": "integer"
          },
          "triggerCount": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "HoldingResponse": {
        "properties": {
          "accountName": {
            "type": "string"
          },
          "accountNumberMasked": {
            "type": "string"
          },
          "assetType": {
            "type": "string"
          },
          "averageCost": {
            "type": "string"
          },
          "classificationReason": {
            "type": "string"
          },
          "costBasis": {
            "type": "string"
          },
          "currentValue": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "lastPrice": {
            "type": "string"
          },
          "quantity": {
            "type": "string"
          },
          "rowNumber": {
            "format": "int32",
            "type": "integer"
          },
          "rowType": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "suggestedClassification": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "ImportAccount": {
        "properties": {
          "accountName": {
            "type": "string"
          },
          "accountNumberMasked": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "PositionResponse": {
        "properties": {
          "accountId": {
            "format": "uuid",
            "type": "string"
          },
          "averageCost": {
            "type": "string"
          },
          "bucket": {
            "type": "string"
          },
          "classification": {
            "type": "string"
          },
          "classificationConfirmed": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "marketValue": {
            "type": "string"
          },
          "quantity": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "PreviewResponse": {
        "properties": {
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/ImportAccount"
            },
            "type": "array"
          },
          "batchId": {
            "format": "uuid",
            "type": "string"
          },
          "cash": {
            "items": {
              "$ref": "#/components/schemas/CashResponse"
            },
            "type": "array"
          },
          "dataAsOf": {
            "format": "date-time",
            "type": "string"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "holdings": {
            "items": {
              "$ref": "#/components/schemas/HoldingResponse"
            },
            "type": "array"
          },
          "status": {
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/SummaryResponse"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "SummaryResponse": {
        "properties": {
          "emergencyCashTarget": {
            "type": "string"
          },
          "errorRowCount": {
            "format": "int32",
            "type": "integer"
          },
          "estimatedCashValue": {
            "type": "string"
          },
          "estimatedInvestedValue": {
            "type": "string"
          },
          "rowCount": {
            "format": "int32",
            "type": "integer"
          },
          "validRowCount": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "emergencyCashTarget",
          "estimatedCashValue",
          "estimatedInvestedValue"
        ],
        "type": "object"
      },
...
      "TradePlanPreviewRequest": {
        "properties": {
          "anchoredToCostBasis": {
            "type": "boolean"
          },
          "averagingDown": {
            "type": "boolean"
          },
          "classification": {
            "minLength": 1,
            "type": "string"
          },
          "classificationConfirmed": {
            "type": "boolean"
          },
          "currentClusterRisk": {
            "maximum": 1,
            "minimum": 0,
            "type": "number"
          },
          "currentOpenStockRisk": {
            "maximum": 1,
            "minimum": 0,
            "type": "number"
          },
          "currentWeight": {
            "maximum": 1,
            "minimum": 0,
            "type": "number"
          },
          "lastDecisionAt": {
            "format": "date-time",
            "type": "string"
          },
          "projectedWeight": {
            "maximum": 1,
            "minimum": 0,
            "type": "number"
          },
          "proposedTradeRisk": {
            "maximum": 1,
            "minimum": 0,
            "type": "number"
          },
          "quality": {
            "minLength": 1,
            "type": "string"
          },
          "thesisImproving": {
            "type": "boolean"
          }
        },
        "required": [
          "classification",
          "currentClusterRisk",
          "currentOpenStockRisk",
          "currentWeight",
          "projectedWeight",
          "proposedTradeRisk",
          "quality"
        ],
        "type": "object"
      },
      "TradePlanPreviewResponse": {
        "properties": {
          "allowed": {
            "type": "boolean"
          },
          "confidence": {
            "type": "string"
          },
          "debug": {
            "type": "boolean"
          },
          "preciseQuantityAllowed": {
            "type": "boolean"
          },
          "projectedClusterRisk": {
            "type": "string"
          },
          "projectedOpenStockRisk": {
            "type": "string"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "risks": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "ruleIds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "tradeRiskCap": {
            "type": "string"
          },
          "weightCap": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Private, manual-execution-only portfolio decision support API",
    "title": "Portfolio Engine API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/etf-dip/preview": {
      "post": {
        "operationId": "preview_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DipPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DipPreviewResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["dip-cashflow-controller"]
      }
    },
...
    "/api/v1/portfolio-imports/fidelity/preview": {
      "post": {
        "operationId": "preview_1",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": ["file"],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["portfolio-import-controller"]
      }
    },
...
    "/api/v1/positions/{id}": {
      "get": {
        "operationId": "position",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["portfolio-controller"]
      }
    },
    "/api/v1/positions/{id}/classification-suggestion": {
      "get": {
        "operationId": "classificationSuggestion",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClassificationSuggestionResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["portfolio-controller"]
      }
    },
    "/api/v1/positions/{id}/classify": {
      "post": {
        "operationId": "classify",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["portfolio-controller"]
      }
    },
...
    "/api/v1/trade-plans/preview": {
      "post": {
        "operationId": "preview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TradePlanPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TradePlanPreviewResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["portfolio-controller"]
      }
    },
...
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost"
    }
  ]
}